+++ cli/tests/test_git_init.rs
@@ -800,6 +800,59 @@ fn test_git_init_colocated_via_flag_git_dir_exists() {
 }
 
 #[test]
+fn test_git_init_colocated_via_config_git_dir_exists() {
+    let test_env = TestEnvironment::default();
+    let work_dir = test_env.work_dir("repo");
+    init_git_repo(work_dir.root(), false);
+
+    test_env.add_config("git.colocate = true");
+
+    let output = test_env.run_jj_in(".", ["git", "init", "repo"]);
+    insta::assert_snapshot!(output, @r#"
+    ------- stderr -------
+    Done importing changes from the underlying Git repo.
+    Initialized repo in "repo"
+    Hint: Running `git clean -xdf` will remove `.jj/`!
+    [EOF]
+    "#);
+
+    // Check that the Git repo's HEAD got checked out
+    insta::assert_snapshot!(get_log_output(&work_dir), @r"
+    @  f3fe58bc88cc
+    ○  e80a42cccd06 my-bookmark git_head() My commit message
+    ◆  000000000000
+    [EOF]
+    ");
+
+    // Check that the Git repo's HEAD moves
+    work_dir.run_jj(["new"]).success();
+    insta::assert_snapshot!(get_log_output(&work_dir), @r"
+    @  0c77f9e21b55
+    ○  f3fe58bc88cc git_head()
+    ○  e80a42cccd06 my-bookmark My commit message
+    ◆  000000000000
+    [EOF]
+    ");
+}
+
+#[test]
+fn test_git_init_no_colocate() {
+    let test_env = TestEnvironment::default();
+    let work_dir = test_env.work_dir("repo");
+
+    test_env.add_config("git.colocate = true");
+
+    let output = test_env.run_jj_in(".", ["git", "init", "--no-colocate", "repo"]);
+    insta::assert_snapshot!(output, @r#"
+    ------- stderr -------
+    Initialized repo in "repo"
+    [EOF]
+    "#);
+
+    assert!(!work_dir.root().join(".git").exists());
+}
+
+#[test]
 fn test_git_init_colocated_via_flag_git_dir_not_exists() {
     let test_env = TestEnvironment::default();
     let work_dir = test_env.work_dir("repo");