+++ lib/src/settings.rs
@@ -63,6 +63,7 @@ pub struct GitSettings {
     pub abandon_unreachable_commits: bool,
     pub executable_path: PathBuf,
     pub write_change_id_header: bool,
+    pub colocate: bool,
 }
 
 impl GitSettings {
@@ -72,6 +73,7 @@ impl GitSettings {
             abandon_unreachable_commits: settings.get_bool("git.abandon-unreachable-commits")?,
             executable_path: settings.get("git.executable-path")?,
             write_change_id_header: settings.get("git.write-change-id-header")?,
+            colocate: settings.get("git.colocate")?,
         })
     }
 }
@@ -83,6 +85,7 @@ impl Default for GitSettings {
             abandon_unreachable_commits: true,
             executable_path: PathBuf::from("git"),
             write_change_id_header: true,
+            colocate: false,
         }
     }
 }